third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-drawImage-video-expected.txt

Ensure correct behavior of drawImage with ImageBitmaps from video elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS Creating a huge ImageBitmap is rejected as expected.
PASS successfullyParsed is true

TEST COMPLETE
